chore: move nb generation script to examples/notebooks/
