+++ .pre-commit-config.yaml
@@ -25,6 +25,10 @@ repos:
     - id: nbqa-black
     - id: nbqa-isort
       args: ["--profile=black"]
+- repo: https://github.com/kynan/nbstripout
+  rev: 0.6.1
+  hooks:
+    - id: nbstripout
 - repo: local
   hooks:
   - id: pytest-check                     # run all tests
