+++ examples/notebooks/README.md
@@ -1,0 +1,2 @@
+> [!IMPORTANT]
+> These notebooks are generated using the [`create_example_nbs.py`](./create_example_nbs.py) script and should not be modified manually. All changes must be made to the respecitve example.
